Fixed Win64 build problems with help of Sameh Khamis
* Added VC10 build files, kindly provided by Sameh Khamis
* Fixed several Win64 related bugs (found by Sameh Khamis):
  - define NAN
  - add some necessary casts of size_t to BigInt
  - replaced occurrence of __PRETTY_FUNCTION__ with FUNCTION_NAME that was forgotten in a previous fix

+++ include/dai/clustergraph.h
@@ -241,7 +241,7 @@ namespace dai {
                     result.insert( Di );
                     if( maxStates ) {
                         totalStates += Di.nrStates();
-                        if( totalStates > maxStates )
+                        if( totalStates > (BigInt)maxStates )
                             DAI_THROW(OUT_OF_MEMORY);
                     }
                     varindices.erase( i );
